METAFONT (x11-Xlib): Avoid gcc warnings by casting screenwidth, screendepth to 'unsigned int'
git-svn-id: svn://tug.org/texlive/trunk/Build/source@32546 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/web2c/window/x11-Xlib.c
@@ -105,7 +105,8 @@ mf_x11_initscreen(void)
     sizehints.height = screendepth;
     sizehints.flags = PPosition|PSize;
 
-    sprintf (default_geometry, "%ux%u+%u+%u", screenwidth, screendepth,
+    sprintf (default_geometry, "%ux%u+%u+%u",
+                               (unsigned int) screenwidth, (unsigned int) screendepth,
                                DEFAULT_X_POSITION, DEFAULT_Y_POSITION);
 
     /* Look up the geometry for this window. (Section 10.2 Obtaining X
